기타 오류 수정

+++ src/member/MemberManager.java
@@ -29,23 +29,45 @@ public class MemberManager {
 	// 사용자에게 Scanner 클래스로 입력 받아 ->
 	// dao ClassMemberDAO 메소드로 저장
 	public void createId() {
+		String mid;
+		String mpw;
+		String mname;
+
 		try {
 			conn = DriverManager.getConnection(jdbcUrl, user, pw);
-
 			System.out.println("회원 가입을 시작합니다.");
+
 			System.out.println("ID를 입력해주세요.");
-			String mid = ir.readString();
-			
-			// 현재 등록하려는 ID가 이미 DB에 있는 경우 종료
-			if (checkDupId(mid)) {
-				System.out.println("이미 존재하는 ID입니다. 다시 입력해주세요.");
+			mid = ir.readString();
+			// ID에 입력하지 않고 enter키 눌렀을 경우.
+			if (checkNullInput(mid)) {
+				System.out.println("ID 생성을 위해 ID를 한 글자 이상 입력해주세요. \n");
 				return;
 			}
-			
+
+			// 현재 등록하려는 ID가 이미 DB에 있는 경우 종료
+			if (checkDupId(mid)) {
+				System.out.println("이미 존재하는 ID입니다. 다시 입력해주세요. \n");
+				return;
+			}
+
 			System.out.println("PW를 입력해주세요.");
-			String mpw = ir.readString();
+			mpw = ir.readString();
+			// PW에 입력하지 않고 enter키 눌렀을 경우.
+			if (checkNullInput(mpw)) {
+				System.out.println("ID 생성을 위해 PW를 한 글자 이상 입력해주세요. \n");
+				return;
+			}
+
 			System.out.println("이름을 입력해주세요.");
-			String mname = ir.readString();
+			mname = ir.readString();
+
+			// 이름에 입력하지 않고 enter키 눌렀을 경우.
+			if (checkNullInput(mname)) {
+				System.out.println("ID 생성을 위해 올바른 이름을 입력해주세요. \n");
+				return;
+			}
+
 			System.out.println("휴대전화번호를 입력해주세요. ex) 010-0000-0000");
 			String mphone = ir.readString();
 			System.out.println("선호하시는 지역을 입력해주세요.");
@@ -70,19 +92,18 @@ public class MemberManager {
 			e.printStackTrace();
 		}
 	}
-	
+
 	private boolean checkDupId(String mid) {
 		ArrayList<String> memberIds = dao.getMemberIds(conn);
-		
-		for(int i = 0; i < memberIds.size(); i++) {
+
+		for (int i = 0; i < memberIds.size(); i++) {
 			if (mid.equals(memberIds.get(i))) {
 				return true;
 			}
 		}
 		return false;
 	}
-	
-	
+
 	// 회원 정보 수정
 	public void editId(String mid) {
 
@@ -92,6 +113,11 @@ public class MemberManager {
 			System.out.println("회원 정보를 수정합니다.");
 			System.out.println("수정하실 PW를 입력해주세요. 변경을 원치 않을 경우 기존 정보를 입력해주세요.");
 			String mpw = ir.readString();
+			// PW에 입력하지 않고 enter키 눌렀을 경우.
+			if (checkNullInput(mpw)) {
+				System.out.println("정보 수정을 위해 PW를 한 글자 이상 입력해주세요.");
+				return;
+			}
 			System.out.println("지역을 입력해주세요. 변경을 원치 않을 경우 기존 정보를 입력해주세요.");
 			String mloc = ir.readString();
 			System.out.println("휴대전화번호를 입력해주세요. ex) 010-0000-0000");
@@ -135,16 +161,27 @@ public class MemberManager {
 	}
 
 	// 회원 탈퇴
-	public void deleteMyId(String mid) {
+	public boolean deleteMyId(String mid) {
 		try {
 			conn = DriverManager.getConnection(jdbcUrl, user, pw);
-
-			dao.deleteId(conn, member);
-			System.out.println("회원 정보가 삭제되었습니다.");
+			System.out.println("ID를 삭제하시겠습니까? (Y/N)");
+			String check = ir.readString();
+			if (check.equals("N") || check.equals("n")) {
+				System.out.println("이전 메뉴로 돌아갑니다.");
+				return true;
+			} else if (check.equals("Y") || check.equals("y")) {
+				dao.deleteId(conn, member);
+				System.out.println("회원 정보가 삭제되었습니다.");
+				return false;
+			} else {
+				System.out.println("올바른 문자를 입력해주세요.");
+				return true;
+			}
 
 		} catch (SQLException e) {
 			e.printStackTrace();
 		}
+		return true;
 	}
 
 	public void editPoint(int total) {
@@ -155,5 +192,12 @@ public class MemberManager {
 		} catch (SQLException e) {
 			e.printStackTrace();
 		}
+	}
+
+	private boolean checkNullInput(String str) {
+		if (str.length() == 0) {
+			return true;
+		}
+		return false;
 	}
 }
